+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
 ## v0.3.11 2026-02-14
+### Improvement
+- Updating today or tomorrow charge schedule may trigger charger to reschedule charge.
 - Refactored SolarCharge adding ChargerScheduler class.
 - Moved local only entities from SolarCharge to sc_option_state.
 - Refactored sc_option_state by removing _get_subentry().
 - Refactored ChargeController adding SolarCharge class.
+### Fix
 - Fixed issue with "Reset charge limit and time" button not working.
 
 ## v0.3.10 2026-02-11